Autorizacao - ok

+++ app/Http/Controllers/ProjectController.php
@@ -5,6 +5,7 @@ namespace CodeProject\Http\Controllers;
 use CodeProject\Repositories\ProjectRepository;
 use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -14,23 +15,24 @@ class ProjectController extends Controller
      */
     protected $repository;
     protected $service;
-    
+
     public function __construct(ProjectRepository $repository, ProjectService $service)
     {
         $this->repository = $repository;
         $this->service = $service;
     }
-    
+
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => Authorizer::getResourceOwnerId()]);
+//        return $this->repository->all();
 //        return $this->repository->with(['owner', 'client'])->all();
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -41,11 +43,14 @@ class ProjectController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->repository->find($id);
     }
 
@@ -57,29 +62,56 @@ class ProjectController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->update($request->all(), $id);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        if ($this->checkProjectOwner($id) == false){
+            return ['error' => 'Access forbidden'];
+        }
         try{
             $this->repository->find($id)->delete();
             return response()->json(['Project deletado com sucesso']);
         }catch (\Exception $e) {
             return response()->json(['Project não encontrado']);
         }
+
     }
-    
+
+    public function checkProjectOwner($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    public function checkProjectMember($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    public function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)){
+            return true;
+        }
+        return false;
+    }
+
 }
